refactor: folder structure and typing file
