+++ packages/topo-imagery-common/src/common/aws/aws_credential_source.py
