+++ packages/topo-imagery-common/src/common/aws/tests/aws_helper_test.py
@@ -1,5 +1,5 @@
 from pytest_subtests import SubTests
-from aws.aws_helper import parse_path
+from common.aws.aws_helper import parse_path
 
 
 def test_parse_path_s3(subtests: SubTests) -> None:
