+++ packages/topo-imagery-common/src/common/cli/cli_helper.py
@@ -6,8 +6,8 @@ from typing import Any, NamedTuple
 
 import shapely.geometry
 from linz_logger import get_log
-from datetimes import parse_rfc_3339_date
-from files.fs import read
+from common.datetimes import parse_rfc_3339_date
+from common.files.fs import read
 
 
 class InputParameterError(Exception):
