+++ packages/topo-imagery-common/src/common/files/files_helper.py
