+++ scripts/stac/imagery/create_stac.py
@@ -2,9 +2,9 @@ import json
 import os
 from typing import Any
 
-from files import checksum, fs
-from files.files_helper import get_file_name_from_path
-from files.fs import NoSuchFileError, read
+from common.files import checksum, fs
+from common.files.files_helper import get_file_name_from_path
+from common.files.fs import NoSuchFileError, read
 from linz_logger import get_log
 from shapely.geometry.base import BaseGeometry
