+++ scripts/stac/imagery/tests/create_stac_test.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 from pathlib import Path
 from typing import Any, cast
 
-from datetimes import format_rfc_3339_datetime_string
+from common.datetimes import format_rfc_3339_datetime_string
 from pytest_subtests import SubTests
 
 from scripts.conftest import any_epoch_datetime, any_epoch_datetime_string
