+++ scripts/stac/imagery/tests/generators.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from os import urandom
 from typing import Callable
 
-from files.checksum import multihash_as_hex
+from common.files.checksum import multihash_as_hex
 
 from scripts.stac.imagery.item import STACAsset, STACProcessing, STACProcessingSoftware
