+++ scripts/standardising.py
@@ -7,12 +7,12 @@ from decimal import Decimal
 from functools import partial
 from multiprocessing import Pool
 
-from aws.aws_helper import is_s3
-from cli.cli_helper import TileFiles
-from files.files_helper import ContentType, is_tiff
-from files.fs import exists, read, write, write_all, write_sidecars
+from common.aws.aws_helper import is_s3
+from common.cli.cli_helper import TileFiles
+from common.files.files_helper import ContentType, is_tiff
+from common.files.fs import exists, read, write, write_all, write_sidecars
+from common.log.time_helper import time_in_ms
 from linz_logger import get_log
-from log.time_helper import time_in_ms
 from tifffile import TiffFile
 
 from scripts.gdal.gdal_bands import get_gdal_band_offset
