+++ scripts/translate_ascii.py
@@ -4,11 +4,11 @@ import tempfile
 from functools import partial
 from multiprocessing import Pool
 
-from cli.common_args import CommonArgumentParser
-from files.files_helper import get_file_name_from_path
-from files.fs import read, write_all, write_sidecars
+from common.cli.common_args import CommonArgumentParser
+from common.files.files_helper import get_file_name_from_path
+from common.files.fs import read, write_all, write_sidecars
+from common.log.time_helper import time_in_ms
 from linz_logger import get_log
-from log.time_helper import time_in_ms
 
 from scripts.gdal.gdal_commands import get_ascii_translate_command
 from scripts.gdal.gdal_helper import run_gdal
